requirements.txt
Django==2.1.5
django-uuslug==1.1.8
pkg-resources==0.0.0
python-slugify==2.0.1
pytz==2018.9
Unidecode==1.0.23
